requirements.txt
alembic==1.14.1
Pillow==11.2.1
APScheduler==3.10.4
asyncpg==0.30.0
email-validator==2.2.0
fastapi==0.116.1
httpx==0.28.1
paramiko==3.5.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
psycopg[binary]==3.2.6
pydantic==2.11.1
pydantic-settings==2.8.1
PyJWT==2.10.1
python-multipart==0.0.20
redis==5.2.1
SQLAlchemy==2.0.39
uvicorn[standard]==0.34.0
